Month 14/step 2

+++ src/controllers/cards.ts
@@ -1,7 +1,8 @@
 import { NextFunction, Request, Response } from "express";
 import Errors from "../errors/errors";
 import Card from "../models/card";
-import { CustomRequest } from "../utils/interfaces";
+import { AuthRequest } from '../middlewares/auth';
+import { ERROR } from '../constants/errors';
 
 export const getCards = (req: Request, res: Response, next: NextFunction) => {
   Card.find({})
@@ -13,7 +14,7 @@ export const getCards = (req: Request, res: Response, next: NextFunction) => {
 };
 
 export const postCard = (
-  req: CustomRequest,
+  req: AuthRequest,
   res: Response,
   next: NextFunction
 ) => {
@@ -29,26 +30,38 @@ export const postCard = (
     })
     .catch((err) => {
       if (err.name === "ValidationError") {
-        next(Errors.badRequest());
+        next(Errors.badRequest(ERROR.message.BAD_REQUEST));
       } else {
         next(err);
       }
     });
 };
 
-export const deleteCard = (req: Request, res: Response, next: NextFunction) => {
-  Card.findByIdAndRemove(req.params.cardId)
+export const deleteCard = (
+  req: AuthRequest,
+  res: Response,
+  next: NextFunction
+) => {
+  const userId = req.user?._id;
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        throw Errors.notFoundRequest();
+        throw Errors.notFound(ERROR.message.NOT_FOUND_REQUEST);
       }
+      if (card.owner.toString() !== userId) {
+        throw Errors.forbiddenError();
+      }
+      return Card.findByIdAndRemove(req.params.cardId);
+    })
+    .then((card) => {
       res.send({
         message: "Карточка удалена",
+        data: card,
       });
     })
     .catch((err) => {
       if (err.name === "CastError") {
-        next(Errors.notFoundRequest());
+        next(Errors.badRequest(ERROR.message.INVALID_ID_ERROR));
       } else {
         next(err);
       }
@@ -56,7 +69,7 @@ export const deleteCard = (req: Request, res: Response, next: NextFunction) => {
 };
 
 export const likeCard = (
-  req: CustomRequest,
+  req: AuthRequest,
   res: Response,
   next: NextFunction
 ) => {
@@ -68,7 +81,7 @@ export const likeCard = (
   )
     .then((card) => {
       if (!card) {
-        throw Errors.notFoundRequest();
+        throw Errors.notFound(ERROR.message.NOT_FOUND_REQUEST);
       }
       res.send({
         message: "Лайк поставлен",
@@ -76,7 +89,7 @@ export const likeCard = (
     })
     .catch((err) => {
       if (err.name === "CastError") {
-        next(Errors.badRequest());
+        next(Errors.badRequest(ERROR.message.INVALID_ID_ERROR));
       } else {
         next(err);
       }
@@ -84,7 +97,7 @@ export const likeCard = (
 };
 
 export const dislikeCard = (
-  req: CustomRequest,
+  req: AuthRequest,
   res: Response,
   next: NextFunction
 ) => {
@@ -96,19 +109,18 @@ export const dislikeCard = (
   )
     .then((card) => {
       if (!card) {
-        throw Errors.notFoundRequest();
+        throw Errors.notFound(ERROR.message.NOT_FOUND_REQUEST);
       }
       res.send(
         // card
-        {message: "Лайк удален"}
-        );
+        { message: "Лайк удален" }
+      );
     })
     .catch((err) => {
       if (err.name === "CastError") {
-        next(Errors.badRequest());
+        next(Errors.badRequest(ERROR.message.INVALID_ID_ERROR));
       } else {
         next(err);
       }
     });
 };
-
